14.0 [ADD] hr_attendance_hour_type_report
This module provides a new report in hr_attendance to display the worked
time of employees split in the following categories:

* type of day (weekday, sunday, holiday)
* daytime / nighttime

The information of a day being a holiday comes from the configured
public holidays from the module hr_holidays_public.

This is meant as a helper for people interacting with external payroll systems.